Fix minor bugs, mostly for linkedin
Thanks to Gijs Hillenius for reporting these bugs

Linkedin began appending ^M chars at the end of lines, remove them

Linkedin began having trailing blank lines, remove them

pcsv brings along text properties when parsing a buffer, which end up as
spam in the bbdb file. Functionally does not seem to be a problem, but
the bbdb file is human readable, and this makes it far less so. I don't
know how to strip the text properties from all the strings. We could use
a temp buffer and file, but I expect csv buffers will always have
underlying files, and it's good for them to anyways to diagnose any
problems, so always save and read from the underlying csv file.